% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.slice.R
\name{add.slice}
\alias{add.slice}
\title{add.slice - Function that adds single slices to a pie chart}
\usage{
add.slice(no, r, pos.x, pos.y, nop, density = NULL, colors = c("#e43157",
  "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33", "#a65628", "#f781bf"))
}
\arguments{
\item{no}{Number of the AE for the given patient, determining color of the AE (numeric)}

\item{r}{Radius of the circle}

\item{pos.x}{Center of the circle on x-lab (numeric)}

\item{pos.y}{Center of the circle on y-lab (numeric)}

\item{nop}{Total number of different AEs (numeric)}

\item{density}{Density of slice coloring (numeric)}

\item{colors}{Colors for slices (this should be maximally 8) (character)}
}
\description{
Creates a polygon (base R) representing a slice in a the piechart
}
\keyword{internal}
