% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sldata.R
\docType{data}
\name{example_sldata}
\alias{example_sldata}
\title{Example data set with simulated subject level information}
\format{A data frame with the following variables (in this order):
\describe{
  \item{SUBJIDN}{subject ID}
  \item{TRT01A}{treatment group}
  \item{TRTSDT}{study day of treatment start}
  \item{LVDT}{study day of drop out}
  \item{DTHDT}{study day of death}
  \item{AGE}{subject age}
  \item{SEX}{subject sex}
  \item{REGION}{subject region}
  \item{SAFFN}{safety flag (numeric)}
}}
\usage{
example_sldata
}
\description{
Example data set with simulated subject level information
}
\keyword{datasets}
