% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_data.R
\docType{data}
\name{patient_data}
\alias{patient_data}
\title{Example data set with simulated patient information}
\format{A data frame with the following variables (in this order):
\describe{
  \item{ps}{subject ID}
  \item{treat}{treatment group}
  \item{end}{study day of drop out}
  \item{death}{study day of death}
}}
\usage{
patient_data
}
\description{
Example data set with simulated patient information
}
\keyword{datasets}
