% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart.R
\name{piechart}
\alias{piechart}
\title{piechart - creates piechart of AEs for on patient}
\usage{
piechart(radii, nos, nop = 8, pos.x = 0, pos.y = 0, c.col = "#383838",
  circle.radius = 0.9, cont = 1, density = NULL)
}
\arguments{
\item{radii}{Radii of the single slices added to the pie chart (numeric)}

\item{nos}{Numbers of AEs to be included in pie chart (numeric)}

\item{nop}{Total number of different AEs (numeric)}

\item{pos.x}{Center of the circle on x-lab (numeric)}

\item{pos.y}{Center of the circle on y-lab (numeric)}

\item{c.col}{Colour of the circle (hexadecimal character)}

\item{cont}{Indicator of patient status}

\item{density}{Density of slice coloring (numeric)}
}
\description{
Creates entire pie chart (base R) for one patient
}
\keyword{internal}
