% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecharts.R
\name{piecharts}
\alias{piecharts}
\title{piecharts - function to depict all piecharts in one graphic given data files}
\usage{
piecharts(i, aes, ae_data, d_data, patients, xlines, ylines)
}
\arguments{
\item{i}{Study day (numeric)}

\item{aes}{Adverse event to be included (character)}

\item{ae_data}{Adverse event dataset on day i (data.frame)}

\item{d_data}{Adverse event dataset before day i (data.frame)}

\item{patients}{Dataset with clinical annotation for patients (data.frame)}

\item{xlines}{Positions for x-axis separation lines (numeric)}

\item{ylines}{Positions for y-axis separation lines (numeric)}
}
\description{
Creates base R plot of all piecharts combined
}
\keyword{internal}
